fixing the navigation controller for all the view controllers

+++ WeatherAPIAPP/Models/WeatherModel.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct main: Codable{
-    var temp : Float
+    var temp : Double
 }
 
 struct WeatherModel: Codable{
